+++ .ci/scripts/local_tests.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+# Copyright 2013 The Flutter Authors. All rights reserved.
+# Use of this source code is governed by a BSD-style license that can be
+# found in the LICENSE file.
+
+# TODO(stuartmorgan): Call scripts/local_tests.sh here as part of updating
+# that script to work on Windows.